Project foundations
Get the project standing and working. The basic functionality of the
site works as expected. Now on to adding styles and additional
functionality through making calls to an api to get actual random
quotes. This was actually kind of fun when I figured out what the hell I
was doing.